+++ typespec-extension/changelog.md
@@ -1,5 +1,11 @@
 # Release History
 
+## 0.13.1 (Unreleased)
+
+Compatible with compiler 0.52.
+
+- Behavior changed on "multipart/form-data" request. The file field would take a `##FileDetails` model, instead of `BinaryData`. 
+
 ## 0.13.0 (2024-01-25)
 
 Compatible with compiler 0.52.